requirements.txt
pandas
numpy
Pillow
matplotlib
tensorflow
scikit-image